refactor : getById 파라미터 수정(int -> long)

+++ src/main/java/roomescape/dao/ReservationJdbcDAO.java
@@ -67,7 +67,7 @@ public class ReservationJdbcDAO implements ReservationDAO {
     }
 
     @Override
-    public Reservation getById(int id) {
+    public Reservation getById(long id) {
         String sql = "SELECT * FROM reservation WHERE id = ?";
 
         try {
